# local/frontend/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ContextSnapshotInput": {
        "additionalProperties": false,
        "description": "Fixed evidence-time context dimensions.\n\n``freshness_basis`` is intentionally an opaque, non-empty source-provided\nstring because the specification requires the basis to be preserved but does\nnot define a controlled vocabulary or timestamp-only representation.",
        "properties": {
          "environment": {
            "$ref": "#/components/schemas/ContextDimensionInput"
          },
          "location": {
            "$ref": "#/components/schemas/ContextDimensionInput"
          },
          "machine_operating_state": {
            "$ref": "#/components/schemas/ContextDimensionInput"
          },
          "shift": {
            "$ref": "#/components/schemas/ContextDimensionInput"
          },
          "workload": {
            "$ref": "#/components/schemas/ContextDimensionInput"
          }
        },
        "required": [
          "shift",
          "location",
          "machine_operating_state",
          "workload",
          "environment"
        ],
        "title": "ContextSnapshotInput",
        "type": "object"
      },
...
      "EvidenceAttachmentInput": {
        "additionalProperties": false,
        "properties": {
          "attachment_type": {
            "minLength": 1,
            "title": "Attachment Type",
            "type": "string"
          },
          "checksum": {
            "minLength": 1,
            "title": "Checksum",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "file_size": {
            "minimum": 0.0,
            "title": "File Size",
            "type": "integer"
          },
          "mime_type": {
            "minLength": 1,
            "title": "Mime Type",
            "type": "string"
          },
          "storage_reference": {
            "minLength": 1,
            "title": "Storage Reference",
            "type": "string"
          }
        },
        "required": [
          "attachment_type",
          "storage_reference",
          "mime_type",
          "file_size",
          "checksum",
          "created_at"
        ],
        "title": "EvidenceAttachmentInput",
        "type": "object"
      },
...
      "EvidenceBundleCompleteness": {
        "additionalProperties": false,
        "properties": {
          "canonical_readiness_policy": {
            "title": "Canonical Readiness Policy",
            "type": "string"
          },
          "incomplete_sections": {
            "items": {
              "$ref": "#/components/schemas/EvidenceBundleSection"
            },
            "title": "Incomplete Sections",
            "type": "array"
          },
          "semantic_retrieval": {
            "$ref": "#/components/schemas/SemanticRetrievalMetadata"
          },
          "status_reason": {
            "title": "Status Reason",
            "type": "string"
          },
          "truncated_sections": {
            "items": {
              "$ref": "#/components/schemas/EvidenceBundleSection"
            },
            "title": "Truncated Sections",
            "type": "array"
          }
        },
        "required": [
          "canonical_readiness_policy",
          "status_reason",
          "semantic_retrieval"
        ],
        "title": "EvidenceBundleCompleteness",
        "type": "object"
      },
      "EvidenceBundleResponse": {
        "additionalProperties": false,
        "properties": {
          "completeness": {
            "$ref": "#/components/schemas/EvidenceBundleCompleteness"
          },
          "evidence_time_context_snapshots": {
            "items": {
              "$ref": "#/components/schemas/EvidenceTimeContextItem"
            },
            "title": "Evidence Time Context Snapshots",
            "type": "array"
          },
          "incident_id": {
            "format": "uuid",
            "title": "Incident Id",
            "type": "string"
          },
          "primary_incident_evidence": {
            "items": {
              "$ref": "#/components/schemas/BundleEvidenceItem"
            },
            "title": "Primary Incident Evidence",
            "type": "array"
          },
          "provenance_index": {
            "items": {
              "$ref": "#/components/schemas/ProvenanceIndexItem"
            },
            "title": "Provenance Index",
            "type": "array"
          },
          "selected_exact_history": {
            "items": {
              "$ref": "#/components/schemas/BundleEvidenceItem"
            },
            "title": "Selected Exact History",
            "type": "array"
          },
          "selected_semantic_history": {
            "items": {
              "$ref": "#/components/schemas/BundleEvidenceItem"
            },
            "title": "Selected Semantic History",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/EvidenceBundleStatus"
          },
          "verification_context": {
            "items": {
              "$ref": "#/components/schemas/VerificationContextItem"
            },
            "title": "Verification Context",
            "type": "array"
          }
        },
        "required": [
          "incident_id",
          "status",
          "primary_incident_evidence",
          "selected_exact_history",
          "selected_semantic_history",
          "verification_context",
          "evidence_time_context_snapshots",
          "provenance_index",
          "completeness"
        ],
        "title": "EvidenceBundleResponse",
        "type": "object"
      },
...
      "EvidenceBundleStatus": {
        "enum": [
          "READY",
          "PARTIAL",
          "INSUFFICIENT_EVIDENCE"
        ],
        "title": "EvidenceBundleStatus",
        "type": "string"
      },
      "EvidenceIngestionResponse": {
        "additionalProperties": false,
        "properties": {
          "evidence_id": {
            "format": "uuid",
            "title": "Evidence Id",
            "type": "string"
          },
          "idempotent_replay": {
            "title": "Idempotent Replay",
            "type": "boolean"
          },
          "session_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          }
        },
        "required": [
          "evidence_id",
          "idempotent_replay"
        ],
        "title": "EvidenceIngestionResponse",
        "type": "object"
      },
...
      "EvidenceTimeContextItem": {
        "additionalProperties": false,
        "properties": {
          "context_snapshots": {
            "items": {
              "$ref": "#/components/schemas/ContextSnapshotOutput"
            },
            "title": "Context Snapshots",
            "type": "array"
          },
          "evidence_id": {
            "format": "uuid",
            "title": "Evidence Id",
            "type": "string"
          },
          "source_classification": {
            "$ref": "#/components/schemas/EvidenceBundleSourceClassification"
          }
        },
        "required": [
          "evidence_id",
          "source_classification"
        ],
        "title": "EvidenceTimeContextItem",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HumanObservationInput": {
        "additionalProperties": false,
        "properties": {
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/EvidenceAttachmentInput"
            },
            "title": "Attachments",
            "type": "array"
          },
          "component_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Component Id"
          },
          "context_snapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContextSnapshotInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "machine_id": {
            "format": "uuid",
            "title": "Machine Id",
            "type": "string"
          },
          "observation_type": {
            "minLength": 1,
            "title": "Observation Type",
            "type": "string"
          },
          "original_source_record_id": {
            "minLength": 1,
            "title": "Original Source Record Id",
            "type": "string"
          },
          "original_timestamp": {
            "format": "date-time",
            "title": "Original Timestamp",
            "type": "string"
          },
          "payload": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Payload",
            "type": "object"
          },
          "provenance": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Provenance",
            "type": "object"
          },
          "raw_payload": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Raw Payload",
            "type": "object"
          },
          "session_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          }
        },
        "required": [
          "machine_id",
          "original_source_record_id",
          "original_timestamp",
          "observation_type",
          "payload",
          "raw_payload",
          "provenance"
        ],
        "title": "HumanObservationInput",
        "type": "object"
      },
...
      "IncidentCollectionResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/IncidentSummaryResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ],
        "title": "IncidentCollectionResponse",
        "type": "object"
      },
...
      "IncidentStatus": {
        "enum": [
          "OPEN",
          "VERIFYING",
          "VERIFIED",
          "RECURRED"
        ],
        "title": "IncidentStatus",
        "type": "string"
      },
      "IncidentSummaryResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "due_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due State"
          },
          "due_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Time"
          },
          "incident_id": {
            "format": "uuid",
            "title": "Incident Id",
            "type": "string"
          },
          "machine_id": {
            "format": "uuid",
            "title": "Machine Id",
            "type": "string"
          },
          "owner_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Ref"
          },
          "severity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity"
          },
          "status": {
            "$ref": "#/components/schemas/IncidentStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "incident_id",
          "machine_id",
          "status",
          "owner_ref",
          "severity",
          "due_state",
          "due_time",
          "created_at",
          "updated_at"
        ],
        "title": "IncidentSummaryResponse",
        "type": "object"
      },
      "JsonValue": {},
...
      "MachineEventInput": {
        "additionalProperties": false,
        "properties": {
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/EvidenceAttachmentInput"
            },
            "title": "Attachments",
            "type": "array"
          },
          "component_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Component Id"
          },
          "context_snapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContextSnapshotInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "event_type": {
            "minLength": 1,
            "title": "Event Type",
            "type": "string"
          },
          "machine_id": {
            "format": "uuid",
            "title": "Machine Id",
            "type": "string"
          },
          "original_source_record_id": {
            "minLength": 1,
            "title": "Original Source Record Id",
            "type": "string"
          },
          "original_timestamp": {
            "format": "date-time",
            "title": "Original Timestamp",
            "type": "string"
          },
          "payload": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Payload",
            "type": "object"
          },
          "provenance": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Provenance",
            "type": "object"
          },
          "raw_payload": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Raw Payload",
            "type": "object"
          },
          "session_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          }
        },
        "required": [
          "machine_id",
          "original_source_record_id",
          "original_timestamp",
          "event_type",
          "payload",
          "raw_payload",
          "provenance"
        ],
        "title": "MachineEventInput",
        "type": "object"
      },
...
      "MaintenanceRecordInput": {
        "additionalProperties": false,
        "properties": {
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/EvidenceAttachmentInput"
            },
            "title": "Attachments",
            "type": "array"
          },
          "component_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Component Id"
          },
          "context_snapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContextSnapshotInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "machine_id": {
            "format": "uuid",
            "title": "Machine Id",
            "type": "string"
          },
          "original_source_record_id": {
            "minLength": 1,
            "title": "Original Source Record Id",
            "type": "string"
          },
          "original_timestamp": {
            "format": "date-time",
            "title": "Original Timestamp",
            "type": "string"
          },
          "payload": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Payload",
            "type": "object"
          },
          "provenance": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Provenance",
            "type": "object"
          },
          "raw_payload": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Raw Payload",
            "type": "object"
          },
          "record_type": {
            "minLength": 1,
            "title": "Record Type",
            "type": "string"
          },
          "session_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          }
        },
        "required": [
          "machine_id",
          "original_source_record_id",
          "original_timestamp",
          "record_type",
          "payload",
          "raw_payload",
          "provenance"
        ],
        "title": "MaintenanceRecordInput",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "MINE-TRACE Backend",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/evidence/human-observations": {
      "post": {
        "operationId": "ingest_human_observation_api_v1_evidence_human_observations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HumanObservationInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceIngestionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Human Observation",
        "tags": [
          "Evidence"
        ]
      }
    },
    "/api/v1/evidence/machine-events": {
      "post": {
        "operationId": "ingest_machine_event_api_v1_evidence_machine_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MachineEventInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceIngestionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Machine Event",
        "tags": [
          "Evidence"
        ]
      }
    },
    "/api/v1/evidence/maintenance-records": {
      "post": {
        "operationId": "ingest_maintenance_record_api_v1_evidence_maintenance_records_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaintenanceRecordInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceIngestionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Maintenance Record",
        "tags": [
          "Evidence"
        ]
      }
    },
...
    "/api/v1/incidents": {
      "get": {
        "operationId": "list_incidents_api_v1_incidents_get",
        "parameters": [
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "machine_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Machine Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/IncidentStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "severity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "in": "query",
            "name": "owner_ref",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner Ref"
            }
          },
          {
            "in": "query",
            "name": "due_state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Due State"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentCollectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Incidents",
        "tags": [
          "Incidents"
        ]
      }
    },
...
    "/api/v1/incidents/{incident_id}/evidence-bundle": {
      "get": {
        "operationId": "get_incident_evidence_bundle_api_v1_incidents__incident_id__evidence_bundle_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceBundleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Incident Evidence Bundle",
        "tags": [
          "Evidence Bundle"
        ]
      }
    },
...
  },
...
}